+++ tests/conftest.py
@@ -22,6 +22,7 @@ from sagemaker import Session
 from sagemaker.local import LocalSession
 from sagemaker.chainer import Chainer
 from sagemaker.mxnet import MXNet
+from sagemaker.rl import RLEstimator
 from sagemaker.pytorch.defaults import PYTORCH_VERSION
 from sagemaker.sklearn.defaults import SKLEARN_VERSION
 from sagemaker.tensorflow.defaults import TF_VERSION
@@ -37,6 +38,8 @@ def pytest_addoption(parser):
     parser.addoption('--chainer-full-version', action='store', default=Chainer.LATEST_VERSION)
     parser.addoption('--mxnet-full-version', action='store', default=MXNet.LATEST_VERSION)
     parser.addoption('--pytorch-full-version', action='store', default=PYTORCH_VERSION)
+    parser.addoption('--rl-coach-full-version', action='store', default=RLEstimator.COACH_LATEST_VERSION)
+    parser.addoption('--rl-ray-full-version', action='store', default=RLEstimator.RAY_LATEST_VERSION)
     parser.addoption('--sklearn-full-version', action='store', default=SKLEARN_VERSION)
     parser.addoption('--tf-full-version', action='store', default=TF_VERSION)
 
@@ -92,12 +95,12 @@ def mxnet_version(request):
     return request.param
 
 
-@pytest.fixture(scope='module', params=["0.4", "0.4.0"])
+@pytest.fixture(scope='module', params=['0.4', '0.4.0'])
 def pytorch_version(request):
     return request.param
 
 
-@pytest.fixture(scope='module', params=["0.20.0"])
+@pytest.fixture(scope='module', params=['0.20.0'])
 def sklearn_version(request):
     return request.param
 
@@ -106,6 +109,16 @@ def sklearn_version(request):
                                         '1.7', '1.7.0', '1.8', '1.8.0', '1.9', '1.9.0',
                                         '1.10', '1.10.0', '1.11', '1.11.0'])
 def tf_version(request):
+    return request.param
+
+
+@pytest.fixture(scope='module', params=['0.11', '0.11.0'])
+def rl_coach_version(request):
+    return request.param
+
+
+@pytest.fixture(scope='module', params=['0.5', '0.5.3'])
+def rl_ray_version(request):
     return request.param
 
 
@@ -122,6 +135,16 @@ def mxnet_full_version(request):
 @pytest.fixture(scope='module')
 def pytorch_full_version(request):
     return request.config.getoption('--pytorch-full-version')
+
+
+@pytest.fixture(scope='module')
+def rl_coach_full_version(request):
+    return request.config.getoption('--rl-coach-full-version')
+
+
+@pytest.fixture(scope='module')
+def rl_ray_full_version(request):
+    return request.config.getoption('--rl-ray-full-version')
 
 
 @pytest.fixture(scope='module')